+++ src/lm_mcmc.c
@@ -115,7 +115,7 @@ SEXP mcmc_new(SEXP Y, SEXP X, SEXP Rweights, SEXP Rprobinit, SEXP Rmodeldim,
 
 
 	prior_m  = compute_prior_probs(model,pmodel,p, modelprior, noInclusionIs1);
-	if (prior_m == 0.0)  Rprintf("warning initial model has 0 prior probabilty\n");
+	if (prior_m == 0.0)  error("initial model has 0 prior probabilty\n");
 	SetModel2(logmargy, shrinkage_m, prior_m, sampleprobs, logmarg, shrinkage, priorprobs, m);
 	SetModel(Rcoef_m, Rse_m, Rmodel_m, mse_m, R2_m,	beta, se, modelspace, mse, R2, m);